requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.3
psycopg2-binary>=2.9
django-seed>=0.3
